Extracted ProtocolVersion
