+++ rosenpass/src/protocol/build_crypto_server.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 use rosenpass_util::mem::{DiscardResultExt, SwapWithDefaultExt};
 use rosenpass_util::{build::Build, result::ensure_or};
 
-use crate::config::ProtocolVersion;
+use crate::config::protocol_version::ProtocolVersion;
 
 use super::basic_types::{SPk, SSk, SymKey};
 use super::osk_domain_separator::OskDomainSeparator;
@@ -151,7 +151,7 @@ pub struct MissingKeypair;
 /// use rosenpass_util::build::Build;
 /// use rosenpass_secret_memory::secret_policy_use_only_malloc_secrets;
 ///
-/// use rosenpass::config::ProtocolVersion;
+/// use rosenpass::config::protocol_version::ProtocolVersion;
 ///
 /// use rosenpass::protocol::basic_types::{SPk, SymKey};
 /// use rosenpass::protocol::{BuildCryptoServer, Keypair, PeerParams};
@@ -329,7 +329,7 @@ impl BuildCryptoServer {
     /// Adding peers to an existing builder:
     ///
     /// ```rust
-    /// use rosenpass::config::ProtocolVersion;
+    /// use rosenpass::config::protocol_version::ProtocolVersion;
     ///
     /// use rosenpass_util::build::Build;
     /// use rosenpass::protocol::basic_types::{SymKey, SPk};
@@ -403,7 +403,7 @@ impl BuildCryptoServer {
     /// use rosenpass_util::build::Build;
     /// use rosenpass_secret_memory::secret_policy_use_only_malloc_secrets;
     ///
-    /// use rosenpass::config::ProtocolVersion;
+    /// use rosenpass::config::protocol_version::ProtocolVersion;
     /// use rosenpass::hash_domains::protocol;
     ///
     /// use rosenpass::protocol::basic_types::{SymKey, SPk};
