+++ rosenpass/tests/app_server_example.rs
@@ -7,7 +7,10 @@ use rosenpass_util::{file::LoadValueB64, functional::run, mem::DiscardResultExt,
 
 use rosenpass::app_server::{AppServer, AppServerTest, MAX_B64_KEY_SIZE};
 use rosenpass::protocol::basic_types::{SPk, SSk, SymKey};
-use rosenpass::{config::{verbosity::Verbosity, ProtocolVersion}, protocol::osk_domain_separator::OskDomainSeparator};
+use rosenpass::{
+    config::{protocol_version::ProtocolVersion, verbosity::Verbosity},
+    protocol::osk_domain_separator::OskDomainSeparator,
+};
 
 #[test]
 #[cfg_attr(miri, ignore)] // unsupported operation: can't call foreign function `mprotect` on OS `linux`
